create ShareExtensions
